Python

+++ 0x02-python-import_modules/calculator_1.py
@@ -6,4 +6,4 @@ def sub(a, b):
 def mul(a, b):
     return(a * b)
 def div(a, b):
-    return(a / b)
+    return int(a / b)
